+++ src/id_translation/utils/__init__.py
@@ -1,11 +1,14 @@
 """Various supporting functions and classes."""
 
 from ._base_metadata import BaseMetadata
-from ._config_utils import ConfigMetadata
+from ._config_utils import ConfigMetadata, EnvConf, EquivalenceConf, Metaconf
 from ._load_toml import load_toml_file
 
 __all__ = [
     "BaseMetadata",
     "ConfigMetadata",
+    "EnvConf",
+    "EquivalenceConf",
+    "Metaconf",
     "load_toml_file",
 ]